+++ packages/components/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @webbo-ui/button
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @webbo-ui/theme@1.0.6
+
 ## 1.0.6
 
 ### Patch Changes