+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @webbo-ui/input
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @webbo-ui/theme@1.0.6
+  - @webbo-ui/button@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes